Add error handling

+++ backend/controllers/product.controller.js
@@ -50,11 +50,16 @@ export const updateProduct = async (req, res) => {
 //Setting up Delete Endpoints
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({success:false, message: "Invalid Product id"})
+    }
+    
     try{
         await Product.findByIdAndDelete(id);
         res.status(200).json({success:true, message: "Product deleted"}) 
     }catch(error){
-        res.status(404).json({success:false, message: "Product not found"})
+        res.status(500).json({success:false, message: "Server Error"})
         console.log("error in deleting product: ", error.message)
     }
 }
